+++ app/src/main/java/org/breezyweather/common/source/BroadcastSource.kt
@@ -17,6 +17,7 @@
 package org.breezyweather.common.source
 
 import android.content.Context
+import android.os.Bundle
 import breezyweather.domain.location.model.Location
 
 /**
@@ -26,12 +27,11 @@ interface BroadcastSource : Source {
 
     // Make sure to also add it to the Manifest!
     val intentAction: String
-    val intentExtra: String
 
     /**
      * Return null if anything happens and you no longer want to send any data
      */
-    fun getData(
+    fun getExtras(
         context: Context, locations: List<Location>
-    ): String?
+    ): Bundle?
 }